pyproject.toml
[project]
name = "crud-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]
